requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asttokens
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
chromadb==1.3.0
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
comm
dataclasses-json==0.6.7
debugpy
decorator
distro==1.9.0
durationpy==0.10
exceptiongroup
executing
filelock==3.20.0
Flask==3.1.2
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.10.0
firebase_admin==7.1.0
google-auth==2.42.1
google-genai==1.47.0
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==1.0.1
humanfriendly==10.0
idna==3.11
importlib_metadata
importlib_resources==6.5.2
ipykernel
ipython
ipython_pygments_lexers
itsdangerous==2.2.0
jedi
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client
jupyter_core
kubernetes==34.1.0
langchain==0.3.27
langchain-chroma==1.0.0
langchain-classic==1.0.0
langchain-community==0.3.31
langchain-core==1.0.2
langchain-experimental==0.3.4
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.38
lark==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nest_asyncio
numpy==2.3.4
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.6.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.11.0
overrides==7.7.0
packaging
pandas==2.3.3
parso
pickleshare
platformdirs
posthog==5.4.0
prompt_toolkit
propcache==0.4.1
protobuf==6.33.0
psutil
pure_eval
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.17.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil
python-dotenv==1.2.1
pytz==2025.2
pywin32
PyYAML==6.0.3
pyzmq
referencing==0.37.0
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.44
stack_data
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
tornado
tqdm==4.67.1
traitlets
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.10.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.38.0
watchfiles==1.1.1
wcwidth
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.22.0
zipp
zstandard==0.25.0
